backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
numpy
google-genai
chromadb
